change values from uint32_t to uint64_t for energy utilization

+++ src/common/slurm_topology.h
@@ -48,7 +48,7 @@
  *  defined here but is really tree plugin related
 \*****************************************************************************/
 struct switch_record {
-	uint32_t consumed_energy;	/* consumed energy, in joules */
+	uint64_t consumed_energy;	/* consumed energy, in joules */
 	int level;			/* level in hierarchy, leaf=0 */
 	uint32_t link_speed;		/* link speed, arbitrary units */
 	char *name;			/* switch name */
