+++ src/plugins/ext_sensors/rrd/ext_sensors_rrd.h
@@ -80,7 +80,7 @@ extern int ext_sensors_p_get_stependdata(struct step_record *step_rec);
 extern List ext_sensors_p_get_config(void);
 
 /* consolidate RRD data */
-extern uint32_t RRD_consolidate(time_t step_starttime, time_t step_endtime,
+extern uint64_t RRD_consolidate(time_t step_starttime, time_t step_endtime,
 				bitstr_t* bitmap_of_nodes);
 
 extern int init(void);
